Food-Desc Back End
I added a method to data interaction service to provide the ability to add a product item to the shopping cart list when the button is pressed in my component. I worked out most of the back-end in my component, so that it will display the info for a product item passed to it. Finally, I left some comments in category component's TS file to give an idea of what I need.